vhpi: Add support for getting/putting IndexedNames
This allows getting/putting nested arrays, and individual elements of an
array. The final index calculation is done lazily, although this
could be done eagerly by moving it to init_indexedName.

The tests for indexednames and the various enum types are extended now
that it is easier to access them.